Model training and inference integration
This PR creates a new container that uses XGBoost to train a priority inference model. The training is triggered by clicking on a button on DefectDojo's interface, which causes a file to be created in a shared volume. The inference container is notified of the new file, and trains the XGBoost model. Inferences are built for an individual user, the one logged in when the button on DefectDojo's interface was clicked.

Once the model is trained, the inference container will generate inferences for all other Findings, save it to a file, and call an API endpoing on Dojo to load the inferences on the database.